Windows only : Updater : Allow uninstall of dlls in VLC plugin folder

+++ es-app/src/Win32ApiSystem.cpp
@@ -870,6 +870,44 @@ std::pair<std::string, int> Win32ApiSystem::updateSystem(const std::function<voi
 		Utils::FileSystem::removeFile(zipFile);
 
 		auto files = Utils::FileSystem::getDirContent(path, true, true);
+
+		auto pluginFolder = Utils::FileSystem::getExePath() + "/plugins";
+		for (auto pluginFile : Utils::FileSystem::getDirContent(pluginFolder, true))
+		{
+			if (Utils::FileSystem::isDirectory(pluginFile))
+				continue;
+
+			Utils::FileSystem::removeFile(pluginFile + ".old");
+
+			std::string pluginRelativeFile = Utils::FileSystem::createRelativePath(pluginFile, pluginFolder, false);
+			if (Utils::String::startsWith(pluginRelativeFile, "./"))
+				pluginRelativeFile = pluginRelativeFile.substr(2);
+
+			bool existsInArchive = false;
+
+			for (auto installedFile : files)
+			{
+				if (Utils::FileSystem::isDirectory(installedFile))
+					continue;
+
+				std::string relative = Utils::FileSystem::createRelativePath(installedFile, path, false);
+				if (Utils::String::startsWith(relative, "./"))
+					relative = relative.substr(2);
+
+				if (relative == pluginRelativeFile)
+				{
+					existsInArchive = true;
+					break;
+				}
+			}
+
+			if (!existsInArchive)
+			{
+				Utils::FileSystem::removeFile(pluginFile);			
+				rename(pluginFile.c_str(), (pluginFile + ".old").c_str());
+			}
+		}
+		
 		for (auto file : files)
 		{
 			std::string relative = Utils::FileSystem::createRelativePath(file, path, false);
@@ -885,6 +923,11 @@ std::pair<std::string, int> Win32ApiSystem::updateSystem(const std::function<voi
 			}
 			else
 			{
+				// Avoid replacing development exe/lib
+				if ((Utils::String::containsIgnoreCase(localPath, "/RelWithDebInfo/") || Utils::String::containsIgnoreCase(localPath, "/Debug/")) && 
+					(Utils::FileSystem::getExtension(localPath) == ".exe" || Utils::FileSystem::getExtension(localPath) == ".lib"))
+					continue;
+
 				if (Utils::FileSystem::exists(localPath))
 				{
 					Utils::FileSystem::removeFile(localPath + ".old");
